feat: add SetSwapIntent and CreateMessage functions for managing swap intents and message creation
